chore: split up BatchMetadata into smaller components
