+++ web-registry/src/components/molecules/BatchMetadata/BatchMetadata.LinkOrDash.tsx
@@ -1,0 +1,12 @@
+import { LinkWithArrow } from '../../atoms';
+
+export const LinkOrDash = ({
+  href,
+  label,
+}: {
+  href?: string;
+  label?: string;
+}): JSX.Element => {
+  if (!href) return <>-</>;
+  return <LinkWithArrow href={href} label={label || ''} />;
+};
