+++ web-registry/src/components/molecules/index.ts
@@ -1,6 +1,6 @@
 export * from './BackgroundImgSection';
 export * from './BatchInfoGrid';
-export * from './BatchMetadata';
+export * from './BatchMetadata/BatchMetadata';
 export * from './BatchTotalsGrid';
 export * from './EcocreditsSection';
 export * from './EditProjectPageFooter';
